refactor: объединенить файлы компонентов в один, рефакторинг хедера и ингредиентов
