+++ src/components/main/burger-constructor/burger-constructor.js
@@ -1,9 +1,9 @@
 import React from "react";
-import mainStyles from '../main.module.css'
+import constructurStyles from './burger-constructor.module.css'
 export default class BurgerConstructor extends React.Component {
 	render() {
 		return(
-			<div className={mainStyles.constructor}>
+			<div className={constructurStyles.constructor}>
 				<p>Конструктор</p>
 			</div>
 		)
