+++ src/components/main/burger-constructor/burger-constructor.module.css
@@ -1,0 +1,5 @@
+.constructor {
+	display: flex;
+	flex-direction: column;
+	width: 100%;
+}
