+++ src/components/main/burger-ingredients/burger-ingredients.module.css
@@ -1,19 +1,32 @@
+.ingredients {
+	display: flex;
+	flex-direction: column;
+	width: 100%;
+}
+
+.tabs {
+	width: 100%;
+	display: flex;
+}
+
 .rowsContainer {
   max-height: calc(100vh - 88px - 156px - 40px - 24px);
-	/* 88 — header, 156 — margins + title, 40px - first top margin
+  /* 88 — header, 156 — margins + title, 40px - first top margin
 	24 — bottom from frame */
   overflow-y: scroll;
 }
+
 .rowsContainer::-webkit-scrollbar {
-	width: 8px;
-	background: #2F2F37;
+  width: 8px;
+  background: #2f2f37;
 }
 
 .rowsContainer::-webkit-scrollbar-thumb {
-	background: #8585AD;
+  background: #8585ad;
 }
+
 .row:first-of-type {
-	margin-top: 0;
+  margin-top: 0;
 }
 
 .list {
@@ -24,6 +37,7 @@
   column-gap: 24px;
   row-gap: 32px;
 }
+
 .item {
   display: flex;
   flex-direction: column;
@@ -38,10 +52,4 @@
   flex-direction: row;
   align-items: center;
   justify-content: center;
-}
-
-.sum {
-}
-
-.image {
 }
